refactor: 페이지네이션 헬퍼 함수를 utils로 분리
