+++ src/components/Factcheck/History/HistoryPagination.tsx
@@ -10,38 +10,7 @@ import {
 } from "@/components/ui/pagination";
 import { cn } from "@/lib/utils";
 import type { Pagination as PaginationType } from "@/types/factcheck";
-
-const getPageGroup = (currentPage: number, totalPages: number, groupSize: number): number[] => {
-  const groupIndex = Math.floor((currentPage - 1) / groupSize);
-  const start = groupIndex * groupSize + 1;
-  const end = Math.min(start + groupSize - 1, totalPages);
-
-  return Array.from({ length: end - start + 1 }, (_, i) => start + i);
-};
-
-const buildPaginationState = (page: number, totalPages: number, pageGroupSize: number) => {
-  const pageItems = getPageGroup(page, totalPages, pageGroupSize);
-  const currentGroupStart = pageItems[0];
-  const currentGroupEnd = pageItems[pageItems.length - 1] ?? currentGroupStart;
-
-  const isFirstPage = page === 1;
-  const isLastPage = page === totalPages;
-  const isFirstGroup = currentGroupStart === 1;
-  const isLastGroup = currentGroupEnd === totalPages;
-
-  const prevGroupFirstPage = Math.max(1, currentGroupStart - pageGroupSize);
-  const nextGroupFirstPage = Math.min(totalPages, currentGroupEnd + 1);
-
-  return {
-    pageItems,
-    isFirstPage,
-    isLastPage,
-    isFirstGroup,
-    isLastGroup,
-    prevGroupFirstPage,
-    nextGroupFirstPage,
-  };
-};
+import { buildPaginationState } from "@/utils/pagination";
 
 interface HistoryPaginationProps {
   pagination: PaginationType;
@@ -55,7 +24,7 @@ const HistoryPagination = ({ pagination, onPageChange, pageGroupSize }: HistoryP
   if (!totalPages || totalPages <= 1) return null;
 
   const {
-    pageItems,
+    pageNumbers,
     isFirstPage,
     isLastPage,
     isFirstGroup,
@@ -90,7 +59,7 @@ const HistoryPagination = ({ pagination, onPageChange, pageGroupSize }: HistoryP
           />
         </PaginationItem>
 
-        {pageItems.map((item) => (
+        {pageNumbers.map((item) => (
           <PaginationItem key={item}>
             <PaginationLink
               onClick={() => onPageChange(item)}
